// ClientApp/src/app/interfaces/i-express-result.ts
export interface IExpressResult {
    maxKNO3g: number;
    maxConcentrationNinKNO3MgPerLiter: number;
    maxConcentrationKinKNO3MgPerLiter: number;
    optimalKNO3g: number;
    optimalConcentrationNinKNO3MgPerLiter: number;
    optimalConcentrationKinKNO3MgPerLiter: number;

    maxKH2PO4g: number;
    maxConcentrationPinKH2PO4MgPerLiter: number;
    maxConcentrationKinKH2PO4MgPerLiter: number;
    optimalKH2PO4g: number
    optimalConcentrationPinKH2PO4MgPerLiter: number;
    optimalConcentrationKinKH2PO4MgPerLiter: number;

    maxK2SO4g: number;
    maxConcentrationKinK2SO4MgPerLiter: number;
    optimalK2SO4g: number;
    optimalConcentrationKinK2SO4MgPerLiter: number;

    maxMgSO4g: number;
    maxConcentrationMginMgSO4MgPerLiter: number;
    optimalMgSO4g: number;
    optimalConcentrationMginMgSO4MgPerLiter: number;
}    